django-todo-react/backend/requirements.txt

aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
frozenlist==1.3.1
github==1.2.7
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1

multidict==6.0.2
oauthlib==3.2.1
pipenv==2022.10.25
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
yarl==1.8.1


